docs: update Carousel page
