+++ docs/app/docs/components/[slug]/components/carousel/indicator.demo.tsx
@@ -1,0 +1,39 @@
+import { Center } from '@/styled-system/jsx'
+import { Carousel, For } from '@cerberus-design/react'
+import { ITEMS } from './items'
+
+export function IndicatorDemo() {
+  return (
+    <Carousel.Root
+      defaultPage={0}
+      padding="1rem"
+      slideCount={ITEMS.length}
+      slidesPerPage={1}
+      spacing="1rem"
+      showIndicators
+    >
+      <Carousel.ItemGroup>
+        <For each={ITEMS}>
+          {(item, idx) => (
+            <Carousel.Item key={item.id} index={idx} asChild>
+              <Center
+                bgColor="page.bg.100"
+                border="1px solid"
+                borderColor="page.border.200"
+                color="page.text.100"
+                h="16rem"
+                rounded="md"
+                transitionProperty="all"
+                transitionDuration="normal"
+                transitionTimingFunction="cubic-bezier(0.075, 0.82, 0.165, 1)"
+                willChange="z-index, transform, box-shadow"
+              >
+                {item.content}
+              </Center>
+            </Carousel.Item>
+          )}
+        </For>
+      </Carousel.ItemGroup>
+    </Carousel.Root>
+  )
+}
