--- docs/app/docs/components/[slug]/components/carousel/items.json
@@ -1,42 +1,0 @@
-[
-  {
-    "id": "slide-1",
-    "content": "Slide 1"
-  },
-  {
-    "id": "slide-2",
-    "content": "Slide 2"
-  },
-  {
-    "id": "slide-3",
-    "content": "Slide 3"
-  },
-  {
-    "id": "slide-4",
-    "content": "Slide 4"
-  },
-  {
-    "id": "slide-5",
-    "content": "Slide 5"
-  },
-  {
-    "id": "slide-6",
-    "content": "Slide 6"
-  },
-  {
-    "id": "slide-7",
-    "content": "Slide 7"
-  },
-  {
-    "id": "slide-8",
-    "content": "Slide 8"
-  },
-  {
-    "id": "slide-9",
-    "content": "Slide 9"
-  },
-  {
-    "id": "slide-10",
-    "content": "Slide 10"
-  }
-]
