+++ docs/app/docs/components/[slug]/components/carousel/items.ts
@@ -1,0 +1,4 @@
+export const ITEMS = Array.from({ length: 10 }, (_, i) => ({
+  id: `slide-${i}`,
+  content: `Slide ${i}`,
+}))
